refactor(models): keep runtime host within package limit
